+++ RELEASENOTES.rst
@@ -4,4 +4,5 @@ Release notes
 1.0.30
 ------
 
+- CSVReader now supports format functions per column
 - Start of release notes